requirements.txt
Flask
PyGithub
openai
pillow
